Move command line function to an independent class
