+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/client/diffsummary/PinotThirdEyeSummaryClient.java
@@ -1,16 +1,14 @@
 package com.linkedin.thirdeye.client.diffsummary;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.common.base.Preconditions;
+import com.google.common.collect.Lists;
 import com.google.common.collect.Multimap;
-import com.linkedin.thirdeye.dashboard.ThirdEyeDashboardConfiguration;
-import com.linkedin.thirdeye.dashboard.resources.SummaryResource;
-import com.linkedin.thirdeye.dashboard.views.diffsummary.SummaryResponse;
-import com.linkedin.thirdeye.datalayer.util.DaoProviderUtil;
-import com.linkedin.thirdeye.datasource.ThirdEyeCacheRegistry;
-import io.dropwizard.configuration.ConfigurationFactory;
-import io.dropwizard.jackson.Jackson;
-import java.io.File;
+import com.linkedin.thirdeye.datasource.MetricExpression;
+import com.linkedin.thirdeye.datasource.MetricFunction;
+import com.linkedin.thirdeye.datasource.ThirdEyeRequest;
+import com.linkedin.thirdeye.datasource.ThirdEyeRequest.ThirdEyeRequestBuilder;
+import com.linkedin.thirdeye.datasource.ThirdEyeResponse;
+import com.linkedin.thirdeye.datasource.cache.QueryCache;
+import com.linkedin.thirdeye.util.ThirdEyeUtils;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
@@ -18,31 +16,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
-
-import javax.validation.Validation;
-import org.apache.commons.cli.CommandLine;
-import org.apache.commons.cli.CommandLineParser;
-import org.apache.commons.cli.DefaultParser;
-import org.apache.commons.cli.HelpFormatter;
-import org.apache.commons.cli.Option;
-import org.apache.commons.cli.OptionBuilder;
-import org.apache.commons.cli.Options;
-import org.apache.commons.cli.ParseException;
-import org.apache.commons.cli.Parser;
 import org.jfree.util.Log;
 import org.joda.time.DateTime;
-
-import com.google.common.collect.Lists;
-import com.linkedin.thirdeye.datasource.MetricExpression;
-import com.linkedin.thirdeye.datasource.MetricFunction;
-import com.linkedin.thirdeye.datasource.ThirdEyeRequest;
-import com.linkedin.thirdeye.datasource.ThirdEyeResponse;
-import com.linkedin.thirdeye.datasource.ThirdEyeRequest.ThirdEyeRequestBuilder;
-import com.linkedin.thirdeye.datasource.cache.QueryCache;
-import com.linkedin.thirdeye.util.ThirdEyeUtils;
-import org.joda.time.DateTimeZone;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * This class generates query requests to the backend database and retrieve the data for summary algorithm.
@@ -57,9 +32,6 @@ import org.slf4j.LoggerFactory;
  * located together.
  */
 public class PinotThirdEyeSummaryClient implements OLAPDataBaseClient {
-  private static final Logger LOG = LoggerFactory.getLogger(PinotThirdEyeSummaryClient.class);
-  private static final ObjectMapper objectMapper = new ObjectMapper();
-
   private final static DateTime NULL_DATETIME = new DateTime();
   private final static int TIME_OUT_VALUE = 120;
   private final static TimeUnit TIME_OUT_UNIT = TimeUnit.SECONDS;
@@ -279,150 +251,6 @@ public class PinotThirdEyeSummaryClient implements OLAPDataBaseClient {
       for (Map.Entry<String, Double> entry : context.entrySet()) {
         entry.setValue(0d);
       }
-    }
-  }
-
-  public static void main(String[] args) throws Exception {
-    Options options = buildOptions();
-    if (args.length == 0) {
-      HelpFormatter formatter = new HelpFormatter();
-      formatter.printHelp(PinotThirdEyeSummaryClient.class.getSimpleName(), options);
-    } else {
-      CommandLineParser parser = new DefaultParser();
-      CommandLine commandLine = parser.parse(options, args);
-      List<String> argList = commandLine.getArgList();
-      Preconditions.checkArgument(argList.size() > 0, "Please provide config directory as parameter");
-
-      String dataset = commandLine.getOptionValue("dataset");
-      String metricName = commandLine.getOptionValue("metric");
-      String dimensions = commandLine.getOptionValue("dimensions", "");
-      String filterJson = commandLine.getOptionValue("filters", "{}");
-      long currentStart = Long.parseLong(commandLine.getOptionValue("currentStart"));
-      long currentEnd = Long.parseLong(commandLine.getOptionValue("currentEnd"));
-      long baselineStart = Long.parseLong(commandLine.getOptionValue("baselineStart"));
-      long baselineEnd = Long.parseLong(commandLine.getOptionValue("baselineEnd"));
-
-      int summarySize = Integer.parseInt(commandLine.getOptionValue("size", "10"));
-      int topDimensions = Integer.parseInt(commandLine.getOptionValue("topDimensions", "3"));
-      String hierarchiesJson = commandLine.getOptionValue("hierarchies", "[]");
-      boolean oneSideError = commandLine.hasOption("oneSideError");
-      boolean manualOrder = commandLine.hasOption("manual");
-      String dataTimeZoneId = commandLine.getOptionValue("timeZone", DateTimeZone.UTC.getID());
-
-      initThirdEyeEnvironment(argList.get(0));
-
-      SummaryResource summaryResource = new SummaryResource();
-      String summaryResultJsonString;
-      if (manualOrder) {
-        summaryResultJsonString = summaryResource
-            .buildSummaryManualDimensionOrder(dataset, metricName, currentStart, currentEnd, baselineStart, baselineEnd,
-                dimensions, filterJson, summarySize, oneSideError, dataTimeZoneId);
-      } else {
-        summaryResultJsonString = summaryResource
-            .buildSummary(dataset, metricName, currentStart, currentEnd, baselineStart, baselineEnd, dimensions,
-                filterJson, summarySize, topDimensions, hierarchiesJson, oneSideError, dataTimeZoneId);
-      }
-
-      SummaryResponse summaryResponse = objectMapper.readValue(summaryResultJsonString, SummaryResponse.class);
-      LOG.info(summaryResponse.toString());
-    }
-
-    // Force closing the connections to data sources.
-    System.exit(0);
-  }
-
-  private static Options buildOptions() {
-    Options options = new Options();
-
-    Option dataset =
-        OptionBuilder.withArgName("name").hasArg().withDescription("dataset name").isRequired().create("dataset");
-    options.addOption(dataset);
-
-    Option metricName =
-        OptionBuilder.withArgName("name").hasArg().withDescription("metric name").isRequired().create("metric");
-    options.addOption(metricName);
-
-    Option dimensions =
-        OptionBuilder.withArgName("list").hasArg().withDescription("dimension names that are separated by comma")
-            .create("dimensions");
-    options.addOption(dimensions);
-
-    Option filters =
-        OptionBuilder.withArgName("json").hasArg().withDescription("filter to apply on the data cube").create("filters");
-    options.addOption(filters);
-
-
-    Option currentStart =
-        OptionBuilder.withArgName("millis").hasArg().withDescription("current start time incluseive").isRequired()
-            .create("currentStart");
-    options.addOption(currentStart);
-
-    Option currentEnd = OptionBuilder.withArgName("millis").hasArg().withDescription("current end time exclusive").isRequired()
-        .create("currentEnd");
-    options.addOption(currentEnd);
-
-    Option baselineStart =
-        OptionBuilder.withArgName("millis").hasArg().withDescription("baseline start time inclusive").isRequired()
-            .create("baselineStart");
-    options.addOption(baselineStart);
-
-    Option baselineEnd = OptionBuilder.withArgName("millis").hasArg().withDescription("baseline end time exclusive").isRequired()
-        .create("baselineEnd");
-    options.addOption(baselineEnd);
-
-    Option size = OptionBuilder.withArgName("number").hasArg().withDescription("size of summary").create("size");
-    options.addOption(size);
-
-    Option topDimension = OptionBuilder.withArgName("number").hasArg().withDescription("number of top dimensions")
-        .create("topDimensions");
-    options.addOption(topDimension);
-
-    Option hierarchies =
-        OptionBuilder.withArgName("json").hasArg().withDescription("dimension hierarchies").create("hierarchies");
-    options.addOption(hierarchies);
-
-    Option oneSideError =
-        OptionBuilder.withArgName("boolean").withDescription("enable one side error summary")
-            .create("oneSideError");
-    options.addOption(oneSideError);
-
-    Option manualOrder =
-        OptionBuilder.withArgName("boolean").withDescription("use manual dimension order")
-            .create("manual");
-    options.addOption(manualOrder);
-
-    Option dateTimeZone =
-        OptionBuilder.withArgName("id").hasArg().withDescription("time zone id in Joda library").create("timeZone");
-    options.addOption(dateTimeZone);
-
-    return options;
-  }
-
-  private static void initThirdEyeEnvironment(String thirdEyeConfigDir) {
-    System.setProperty("dw.rootDir", thirdEyeConfigDir);
-    String dashboardConfigFilePath = thirdEyeConfigDir + "/dashboard.yml";
-    File configFile = new File(dashboardConfigFilePath);
-
-    ConfigurationFactory<ThirdEyeDashboardConfiguration> factory =
-        new ConfigurationFactory<>(ThirdEyeDashboardConfiguration.class,
-            Validation.buildDefaultValidatorFactory().getValidator(), Jackson.newObjectMapper(), "");
-    ThirdEyeDashboardConfiguration config;
-    try {
-      config = factory.build(configFile);
-    } catch (Exception e) {
-      LOG.error("Exception while constructing ThirdEye config:", e);
-      throw new RuntimeException(e);
-    }
-
-    String persistenceConfig = thirdEyeConfigDir + "/persistence.yml";
-    LOG.info("Loading persistence config from [{}]", persistenceConfig);
-    DaoProviderUtil.init(new File(persistenceConfig));
-
-    try {
-      ThirdEyeCacheRegistry.initializeCachesWithoutRefreshing(config);
-    } catch (Exception e) {
-      LOG.error("Exception while loading caches:", e);
-      throw new RuntimeException(e);
     }
   }
 }
